% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{annotate_seurat}
\alias{annotate_seurat}
\title{Annotate from Seurat object}
\usage{
annotate_seurat(
  seurat_obj,
  tissue,
  api_key = NULL,
  model = "deepseek",
  top_markers = 15
)
}
\arguments{
\item{seurat_obj}{Seurat object with clusters}

\item{tissue}{Tissue name}

\item{api_key}{API key}

\item{model}{Model to use}

\item{top_markers}{Number of top markers per cluster}
}
\value{
Annotation result object
}
\description{
Annotate from Seurat object
}
