% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{are_related_terms}
\alias{are_related_terms}
\title{Check whether two Cell Ontology IDs are lineage-related}
\usage{
are_related_terms(cl_id_1, cl_id_2, ontology)
}
\arguments{
\item{cl_id_1}{First CL identifier.}

\item{cl_id_2}{Second CL identifier.}

\item{ontology}{Ontology object from load_cell_ontology().}
}
\value{
Logical.
}
\description{
Check whether two Cell Ontology IDs are lineage-related
}
