% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{batch_annotate}
\alias{batch_annotate}
\title{Batch annotate multiple samples}
\usage{
batch_annotate(samples, api_key = NULL, model = "deepseek")
}
\arguments{
\item{samples}{List of samples, each with markers and tissue}

\item{api_key}{API key}

\item{model}{Model to use}
}
\value{
List of annotation results
}
\description{
Batch annotate multiple samples
}
