% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{calculate_clade_accuracy}
\alias{calculate_clade_accuracy}
\title{Calculate clade accuracy}
\usage{
calculate_clade_accuracy(predicted_cl, ground_truth_cl, ontology)
}
\arguments{
\item{predicted_cl}{Vector of predicted CL IDs.}

\item{ground_truth_cl}{Vector of ground-truth CL IDs.}

\item{ontology}{Ontology object from load_cell_ontology().}
}
\value{
Proportion of lineage-related predictions.
}
\description{
Calculate clade accuracy
}
