% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_dashboard}
\alias{create_dashboard}
\title{Create summary dashboard}
\usage{
create_dashboard(result)
}
\arguments{
\item{result}{Annotation result object}
}
\value{
Patchwork object with multiple plots
}
\description{
Create summary dashboard
}
