% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_loader.R
\name{get_default_ontology_path}
\alias{get_default_ontology_path}
\title{Get default ontology path}
\usage{
get_default_ontology_path()
}
\value{
Path to cl.obo file
}
\description{
Get default ontology path
}
