% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_model_config}
\alias{get_model_config}
\title{Get model configuration}
\usage{
get_model_config(model_name)
}
\arguments{
\item{model_name}{Model identifier.}
}
\value{
Model configuration list.
}
\description{
Get model configuration
}
