% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{normalize_cell_type}
\alias{normalize_cell_type}
\title{Normalize cell type string for ontology matching}
\usage{
normalize_cell_type(x)
}
\arguments{
\item{x}{Character vector of cell type names.}
}
\value{
Normalized character vector.
}
\description{
Normalize cell type string for ontology matching
}
