% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Create confusion matrix heatmap}
\usage{
plot_confusion_matrix(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Table from calculate_metrics()}
}
\value{
ggplot2 object
}
\description{
Create confusion matrix heatmap
}
