% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ontology_coverage}
\alias{plot_ontology_coverage}
\title{Create ontology coverage plot}
\usage{
plot_ontology_coverage(result)
}
\arguments{
\item{result}{Annotation result object}
}
\value{
ggplot2 object
}
\description{
Create ontology coverage plot
}
