% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{print_annotation_report}
\alias{print_annotation_report}
\title{Print formatted report to console}
\usage{
print_annotation_report(result)
}
\arguments{
\item{result}{Annotation result object}
}
\description{
Print formatted report to console
}
