% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_brain}
\alias{ref_brain}
\title{Reference brain expression data}
\format{
A \code{SummarizedExperiment} object with 21,214 features and 358 samples.
}
\usage{
ref_brain
}
\description{
A bundled reference object used by examples and benchmarking workflows for
brain cell type annotation.
}
\keyword{datasets}
