% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_pancreas}
\alias{ref_pancreas}
\title{Reference pancreas expression data}
\format{
A \code{SingleCellExperiment} object.
}
\usage{
ref_pancreas
}
\description{
A bundled reference object used by examples and benchmarking workflows for
pancreas cell type annotation.
}
\keyword{datasets}
