% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_pbmc}
\alias{ref_pbmc}
\title{Reference peripheral blood mononuclear cell expression data}
\format{
A \code{SummarizedExperiment} object with 46,077 features and 114 samples.
}
\usage{
ref_pbmc
}
\description{
A bundled reference object used by examples and benchmarking workflows for
immune cell type annotation.
}
\keyword{datasets}
