% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{save_report}
\alias{save_report}
\title{Save report to file}
\usage{
save_report(result, filename, format = "txt")
}
\arguments{
\item{result}{Annotation result object}

\item{filename}{Output filename}

\item{format}{Report format ("txt", "json", "yaml")}
}
\description{
Save report to file
}
