% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{seurat_markers_to_list}
\alias{seurat_markers_to_list}
\title{Convert Seurat object markers to DeepSeekCell format}
\usage{
seurat_markers_to_list(seurat_markers, top_n = 15)
}
\arguments{
\item{seurat_markers}{Output from Seurat::FindAllMarkers()}

\item{top_n}{Number of top markers per cluster}
}
\value{
List formatted for annotate_cell_types()
}
\description{
Convert Seurat object markers to DeepSeekCell format
}
