% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{data}
\name{MODELS}
\alias{MODELS}
\title{Unified API Handler for LLM Cell Type Annotation}
\format{
An object of class \code{list} of length 2.
}
\usage{
MODELS
}
\description{
Supports DeepSeek and local Ollama endpoints.
}
\keyword{datasets}
